requirements.txt
Django>=5.0
psycopg2-binary>=2.9
django-environ>=0.11.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-filter>=23.5
drf-yasg>=1.21
pyotp>=2.9
pytest>=8.0
pytest-django>=4.5
pytest-factoryboy>=2.5
factory-boy>=3.2
Pillow>=10.0 
pre-commit>=3.0 
django-cors-headers>=4.0 